requirements.txt
paho_mqtt==1.6.1
HAP-python[QRCode]==4.9.0
HAP-python==4.9.0
python-dotenv==1.0.0
zeroconf==0.118.0
